requirements.txt
pyyaml>=6.0.1
pydantic>=2.8.2
typer>=0.12.3
openai>=1.51.0
rich>=13.8.0
